+++ lib/appium_lib_core/common/base/http_default.rb
@@ -12,12 +12,20 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
+require 'securerandom'
+
 require_relative '../../version'
 
 module Appium
   module Core
     class Base
       module Http
+        module RequestHeaders
+          KEYS = {
+            idempotency: 'X-Idempotency-Key'
+          }.freeze
+        end
+
         class Default < Selenium::WebDriver::Remote::Http::Default
           DEFAULT_HEADERS = {
             'Accept' => CONTENT_TYPE,
@@ -25,6 +33,19 @@ module Appium
             'User-Agent' =>
               "appium/ruby_lib_core/#{VERSION} (#{::Selenium::WebDriver::Remote::Http::Common::DEFAULT_HEADERS['User-Agent']})"
           }.freeze
+
+          attr_accessor :additional_headers
+
+          def initialize(open_timeout: nil, read_timeout: nil, enable_idempotency_header: true)
+            @open_timeout = open_timeout
+            @read_timeout = read_timeout
+
+            @additional_headers = if enable_idempotency_header
+                                    { RequestHeaders::KEYS[:idempotency] => SecureRandom.uuid }
+                                  else
+                                    {}
+                                  end
+          end
 
           # Update <code>server_url</code> provided when ruby_lib _core created a default http client.
           # Set <code>@http</code> as nil to re-create http client for the <code>server_url</code>
@@ -65,18 +86,19 @@ module Appium
           def call(verb, url, command_hash)
             url      = server_url.merge(url) unless url.is_a?(URI)
             headers  = DEFAULT_HEADERS.dup
+            headers  = headers.merge @additional_headers unless @additional_headers.empty?
             headers['Cache-Control'] = 'no-cache' if verb == :get
 
             if command_hash
               payload                   = JSON.generate(command_hash)
               headers['Content-Length'] = payload.bytesize.to_s if [:post, :put].include?(verb)
-
-              ::Appium::Logger.info("   >>> #{url} | #{payload}")
-              ::Appium::Logger.debug("     > #{headers.inspect}")
             elsif verb == :post
               payload = '{}'
               headers['Content-Length'] = '2'
             end
+
+            ::Appium::Logger.info("   >>> #{url} | #{payload}")
+            ::Appium::Logger.info("     > #{headers.inspect}")
 
             request verb, url, headers, payload
           end