+++ CHANGELOG.md
@@ -5,6 +5,8 @@ Read `release_notes.md` for commit level details.
 ## [Unreleased]
 
 ### Enhancements
+- Add `x-idempotency-key` header support (https://github.com/appium/appium-base-driver/pull/400)
+    - Can disable the header with `enable_idempotency_header: false` in `appium_lib` capability. Defaults to `true`.
 
 ### Bug fixes